+++ source/gear/warcry-base.js
@@ -4,7 +4,7 @@ const { joinAsStatement } = require('../util/textUtil');
 
 module.exports = new GearTemplate("War Cry",
 	[
-		["use", "Also target all foes with @{mod0}"],
+		["use", "Stagger a single foe and all foes with @{mod0}"],
 		["Critical💥", "Stagger +@{bonus}"]
 	],
 	"Technique",
@@ -17,14 +17,14 @@ module.exports = new GearTemplate("War Cry",
 			targetSet.add(initialTarget.name);
 			targetArray.push(initialTarget);
 		}
+		const { element, stagger, bonus, modifiers: [targetModifier] } = module.exports;
 		for (const enemy of adventure.room.enemies) {
-			if (enemy.hp > 0 && enemy.getModifierStacks("Exposed") > 0 && !targetSet.has(enemy.name)) {
+			if (enemy.hp > 0 && enemy.getModifierStacks(targetModifier.name) > 0 && !targetSet.has(enemy.name)) {
 				targetSet.add(enemy.name);
 				targetArray.push(enemy);
 			}
 		}
 
-		const { element, stagger, bonus } = module.exports;
 		let pendingStaggerStacks = stagger;
 		if (user.element === element) {
 			pendingStaggerStacks += 2;
@@ -37,7 +37,7 @@ module.exports = new GearTemplate("War Cry",
 	}
 ).setTargetingTags({ type: "single", team: "foe" })
 	.setUpgrades("Charging War Cry", "Slowing War Cry", "Tormenting War Cry")
-	.setModifiers({ name: "Exposed", stacks: 0 })
+	.setModifiers({ name: "Distracted", stacks: 0 })
 	.setStagger(2)
 	.setBonus(2) // Stagger stacks
 	.setDurability(15)