+++ source/gear/warcry-slowing.js
@@ -4,7 +4,7 @@ const { joinAsStatement } = require('../util/textUtil.js');
 
 module.exports = new GearTemplate("Slowing War Cry",
 	[
-		["use", "Inflict @{mod0Stacks} @{mod0} on a foe; also target all foes with @{mod1}"],
+		["use", "Inflict @{mod0Stacks} @{mod0} and Stagger on a foe; also target all foes with @{mod1}"],
 		["Critical💥", "Stagger +@{bonus}"]
 	],
 	"Technique",
@@ -17,14 +17,14 @@ module.exports = new GearTemplate("Slowing War Cry",
 			targetSet.add(initialTarget.name);
 			targetArray.push(initialTarget);
 		}
+		const { element, modifiers: [slow, targetModifier], stagger, bonus } = module.exports;
 		for (const enemy of adventure.room.enemies) {
-			if (enemy.hp > 0 && enemy.getModifierStacks("Exposed") > 0 && !targetSet.has(enemy.name)) {
+			if (enemy.hp > 0 && enemy.getModifierStacks(targetModifier.name) > 0 && !targetSet.has(enemy.name)) {
 				targetSet.add(enemy.name);
 				targetArray.push(enemy);
 			}
 		}
 
-		const { element, modifiers: [slow], stagger, bonus } = module.exports;
 		let pendingStaggerStacks = stagger;
 		if (user.element === element) {
 			pendingStaggerStacks += 2;
@@ -33,11 +33,11 @@ module.exports = new GearTemplate("Slowing War Cry",
 			pendingStaggerStacks += bonus;
 		}
 		changeStagger(targetArray, pendingStaggerStacks);
-		return [joinAsStatement(false, [...targetSet], "was", "were", "Staggered."), ...generateModifierResultLines(combineModifierReceipts(addModifier(targetArray, slow))) ];
+		return [joinAsStatement(false, [...targetSet], "was", "were", "Staggered."), ...generateModifierResultLines(combineModifierReceipts(addModifier(targetArray, slow)))];
 	}
 ).setTargetingTags({ type: "single", team: "foe" })
 	.setSidegrades("Charging War Cry", "Tormenting War Cry")
-	.setModifiers({ name: "Slow", stacks: 1 }, { name: "Exposed", stacks: 0 })
+	.setModifiers({ name: "Slow", stacks: 1 }, { name: "Distracted", stacks: 0 })
 	.setStagger(2)
 	.setBonus(2) // Stagger stacks
 	.setDurability(15)