+++ source/gear/warcry-tormenting.js
@@ -5,7 +5,7 @@ const { isDebuff } = require('../modifiers/_modifierDictionary.js');
 
 module.exports = new GearTemplate("Tormenting War Cry",
 	[
-		["use", "Duplicate debuffs on a foe; also target all foes with @{mod0}"],
+		["use", "Duplicate debuffs on a foe and Stagger them; also target all foes with @{mod0}"],
 		["Critical💥", "Stagger +@{bonus}"]
 	],
 	"Technique",
@@ -18,14 +18,14 @@ module.exports = new GearTemplate("Tormenting War Cry",
 			targetSet.add(initialTarget.name);
 			targetArray.push(initialTarget);
 		}
+		const { element, stagger, bonus, modifiers: [targetModifier] } = module.exports;
 		for (const enemy of adventure.room.enemies) {
-			if (enemy.hp > 0 && enemy.getModifierStacks("Exposed") > 0 && !targetSet.has(enemy.name)) {
+			if (enemy.hp > 0 && enemy.getModifierStacks(targetModifier.name) > 0 && !targetSet.has(enemy.name)) {
 				targetSet.add(enemy.name);
 				targetArray.push(enemy);
 			}
 		}
 
-		const { element, stagger, bonus } = module.exports;
 		let pendingStaggerStacks = stagger;
 		if (user.element === element) {
 			pendingStaggerStacks += 2;
@@ -47,7 +47,7 @@ module.exports = new GearTemplate("Tormenting War Cry",
 	}
 ).setTargetingTags({ type: "single", team: "foe" })
 	.setSidegrades("Charging War Cry", "Slowing War Cry")
-	.setModifiers({ name: "Exposed", stacks: 0 })
+	.setModifiers({ name: "Distracted", stacks: 0 })
 	.setStagger(2)
 	.setBonus(2) // Stagger stacks
 	.setDurability(15)